Feat : Reservation

+++ src/main/java/com/innovationcamp/messenger/domain/wallet/service/GroupWalletService.java
@@ -5,15 +5,15 @@ import com.innovationcamp.messenger.domain.channel.repository.ChannelRepository;
 import com.innovationcamp.messenger.domain.channel.repository.UserChannelRepository;
 import com.innovationcamp.messenger.domain.user.entity.User;
 import com.innovationcamp.messenger.domain.wallet.config.WalletPasswordEncoder;
-import com.innovationcamp.messenger.domain.wallet.dto.GroupWalletCreateDto;
-import com.innovationcamp.messenger.domain.wallet.dto.GroupWalletResponseDto;
-import com.innovationcamp.messenger.domain.wallet.dto.TransactionResponseDto;
-import com.innovationcamp.messenger.domain.wallet.dto.WalletUserResponseDto;
-import com.innovationcamp.messenger.domain.wallet.entity.UserAuthorityEnum;
+import com.innovationcamp.messenger.domain.wallet.dto.*;
 import com.innovationcamp.messenger.domain.wallet.entity.GroupWallet;
-import com.innovationcamp.messenger.domain.wallet.entity.UserGroupWallet;
 import com.innovationcamp.messenger.domain.wallet.entity.Transaction;
-import com.innovationcamp.messenger.domain.wallet.repository.*;
+import com.innovationcamp.messenger.domain.wallet.entity.UserAuthorityEnum;
+import com.innovationcamp.messenger.domain.wallet.entity.UserGroupWallet;
+import com.innovationcamp.messenger.domain.wallet.repository.GroupWalletRepository;
+import com.innovationcamp.messenger.domain.wallet.repository.ReservationRepository;
+import com.innovationcamp.messenger.domain.wallet.repository.TransactionRepository;
+import com.innovationcamp.messenger.domain.wallet.repository.UserGroupWalletRepository;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -37,6 +37,8 @@ public class GroupWalletService {
     private ChannelRepository channelRepository;
     @NonNull
     private UserChannelRepository userChannelRepository;
+    @NonNull
+    private ReservationRepository reservationRepository;
 
     private final Long money = 0L;
 
@@ -49,8 +51,14 @@ public class GroupWalletService {
             throw new IllegalArgumentException("채널의 구성원이 아닙니다.");
     }
 
+    private void validateUserGroupWallet(Long groupWalletId, Long userId){
+        if(!userGroupWalletRepository.existsByGroupWalletIdAndUserId(groupWalletId, userId)){
+            throw new IllegalArgumentException("그룹통장의 구성원이 아닙니다.");
+        }
+    }
+
     @Transactional
-    public GroupWallet createGroupWallet(User user, GroupWalletCreateDto requestDto) {
+    public GroupWalletResponseDto createGroupWallet(User user, GroupWalletCreateDto requestDto) {
         Channel channel = channelRepository.findById(requestDto.getChannelId()).orElseThrow(() -> new IllegalArgumentException("없는 채널입니다."));
         validateUserChannel(requestDto.getChannelId(), user.getId());
         GroupWallet groupWallet = GroupWallet.builder()
@@ -65,26 +73,29 @@ public class GroupWalletService {
 
         UserGroupWallet userGroupWallet = new UserGroupWallet(UserAuthorityEnum.ADMIN, user, groupWallet);
         userGroupWalletRepository.save(userGroupWallet);
-        return groupWallet;
+        return new GroupWalletResponseDto(groupWallet);
     }
 
-    public GroupWallet getGroupWalletById(User user, Long groupWalletId) {
-        validateUserChannel(groupWalletId, user.getId());
-        return findGroupWalletById(groupWalletId);
+    public GroupWalletResponseDto getGroupWalletById(User user, Long groupWalletId) {
+        validateUserGroupWallet(groupWalletId, user.getId());
+        return new GroupWalletResponseDto(findGroupWalletById(groupWalletId));
     }
 
     @Transactional
-    public GroupWallet deleteGroupWalletById(Long groupWalletId) {
+    public String  deleteGroupWalletById(User user, Long groupWalletId) {
         GroupWallet groupWallet = findGroupWalletById(groupWalletId);
+        UserGroupWallet userGroupWallet = userGroupWalletRepository.findByUserAndGroupWallet(user, groupWallet)
+                .orElseThrow(() -> new IllegalArgumentException("groupWallet 에 참여하지 않은 user 입니다."));
+        if (!userGroupWallet.getUserAuthority().equals(UserAuthorityEnum.ADMIN))
+            throw new IllegalArgumentException("groupWallet 에 권한이 없는 user 입니다.");
+
         userGroupWalletRepository.deleteByGroupWallet(groupWallet);
         groupWalletRepository.delete(groupWallet);
-        return groupWallet;
+        return "그룹통장이 해지되었습니다.";
     }
 
     public List<TransactionResponseDto> getTransactionByGroupWallet(User user, Long groupWalletId) {
-        if(!userGroupWalletRepository.existsByGroupWalletIdAndUserId(groupWalletId, user.getId())){
-            throw new IllegalArgumentException("참가하지 않은 Group 통장입니다.");
-        }
+        validateUserGroupWallet(groupWalletId, user.getId());
         GroupWallet groupWallet = findGroupWalletById(groupWalletId);
         List<Transaction> transactionList = transactionRepository.findAllByWallet(groupWallet);
         return transactionList.stream().map(TransactionResponseDto::new).collect(Collectors.toList());
@@ -96,7 +107,7 @@ public class GroupWalletService {
         return userGroupWalletList.stream().map(w -> new WalletUserResponseDto(w.getUser(), w.getUserAuthority())).collect(Collectors.toList());
     }
 
-    public GroupWallet participantGroupWalletById(User user, Long groupWalletId) {
+    public GroupWalletResponseDto participantGroupWalletById(User user, Long groupWalletId) {
         GroupWallet groupWallet = findGroupWalletById(groupWalletId);
         validateUserChannel(groupWallet.getChannel().getId(), user.getId());
         UserGroupWallet userGroupWallet = userGroupWalletRepository.findByUserAndGroupWallet(user, groupWallet).orElse(null);
@@ -104,12 +115,18 @@ public class GroupWalletService {
         //일단 권한 USER
         userGroupWallet = new UserGroupWallet(UserAuthorityEnum.USER, user, groupWallet);
         userGroupWalletRepository.save(userGroupWallet);
-        return groupWallet;
+        return new GroupWalletResponseDto(groupWallet);
     }
 
     public List<GroupWalletResponseDto> getAllGroupWalletByChannelId(User user, Long channelId) {
         Channel channel = channelRepository.findById(channelId).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 채널입니다."));
         validateUserChannel(channelId, user.getId());
         return groupWalletRepository.findAllByChannel(channel).stream().map(GroupWalletResponseDto::new).collect(Collectors.toList());
+    }
+
+    public List<ReservationResponseDto> getAllReservationByGroupWallet(User user, Long groupWalletId) {
+        GroupWallet groupWallet = findGroupWalletById(groupWalletId);
+        validateUserGroupWallet(groupWalletId, user.getId());
+        return reservationRepository.findAllByWalletId(groupWallet.getId()).stream().map(ReservationResponseDto::new).toList();
     }
 }
